[Unified Tabstrip] Fix unified tab group header style.
Since the sizing differ between vertical and horizontal, move insets and
radius to styles class. Add calls to get sizing based on orientation,
and make sizing information static.

Bug: 547962975
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8346508
Commit-Queue: Vince Lugli <lugli@google.com>
Cr-Commit-Position: refs/heads/main@{#1694769}

+++ chrome/browser/ui/views/tabs/tab_group_header.cc
@@ -758,9 +758,9 @@ void TabGroupHeader::CreateHeaderWithTitle() {
       sync_icon_width + padding_between_label_sync_icon + text_width +
       attention_indicator_width + attention_indicator_padding;
   const gfx::Insets title_chip_insets = group_style_->GetInsetsForHeaderChip();
-  const int title_chip_width = std::max(
-      group_style_->GetEmptyChipSize(),
-      static_cast<float>(total_content_width + title_chip_insets.width()));
+  const int title_chip_width =
+      std::max(group_style_->GetEmptyChipSize(),
+               total_content_width + title_chip_insets.width());
 
   const int chip_height = GetChipHeight();
   const int chip_y = GetChipY();
